refactor: reverse `preset.ts` and `@preset/cli.ts`
